requirements.txt
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.22.0
pandas>=1.4.0
scipy>=1.8.0
scikit-image>=0.19.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.2
opencv-python>=4.6.0
tqdm>=4.64.0
Pillow>=9.2.0
monai>=1.0.0
nibabel>=4.0.0
tensorboard>=2.10.0
imageio>=2.22.0
ipython>=8.4.0 